Move to `@prezly/analytics-nextjs` package
