+++ components/RichText/Link.tsx
@@ -1,7 +1,6 @@
+import { STORY_LINK, useAnalytics } from '@prezly/analytics-nextjs';
 import type { LinkNode } from '@prezly/slate-types';
 import type { ReactNode } from 'react';
-
-import { STORY_LINK, useAnalytics } from '@/modules/analytics';
 
 import styles from './styles.module.scss';
