--- modules/analytics/components/CookieConsentLink/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CookieConsentLink';
