--- modules/analytics/components/index.ts
@@ -1,3 +1,0 @@
-export { default as Analytics } from './Analytics';
-export { default as CookieConsentBar } from './CookieConsentBar';
-export { default as CookieConsentLink } from './CookieConsentLink';
