--- modules/analytics/events/campaign.ts
@@ -1,3 +1,0 @@
-export const CAMPAIGN = {
-    CLICK: 'Campaign Click',
-};
