--- modules/analytics/events/index.ts
@@ -1,2 +1,0 @@
-export * from './campaign';
-export * from './story';
