--- modules/analytics/events/story.ts
@@ -1,31 +1,0 @@
-export const STORY_ASSET = {
-    CLICK: 'Story Asset Click',
-};
-
-export const STORY_CONTACT = {
-    CLICK: 'Story Contact Click',
-};
-
-export const STORY_EMBED = {
-    CLICK: 'Story Embed Click',
-};
-
-export const STORY_FILE = {
-    CLICK: 'Story File Click',
-    DOWNLOAD: 'Story File Download',
-};
-
-export const STORY_GALLERY_IMAGE = {
-    DOWNLOAD: 'Gallery Image Download',
-    VIEW: 'Gallery Image View',
-};
-
-export const STORY_IMAGE = {
-    CLICK: 'Story Image Click',
-    DOWNLOAD: 'Story Image Download',
-    VIEW: 'Story Image View',
-};
-
-export const STORY_LINK = {
-    CLICK: 'Story Link Click',
-};
