--- modules/analytics/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useAnalytics';
-export * from './useCookieConsent';
