--- modules/analytics/index.ts
@@ -1,5 +1,0 @@
-export * from './components';
-export { AnalyticsContextProvider } from './context';
-export * from './events';
-export { useAnalytics } from './hooks';
-export type { AnalyticsJS } from './types';
