--- modules/analytics/lib/index.ts
@@ -1,8 +1,0 @@
-export * from './cookieState';
-export * from './createAnalyticsStub';
-export * from './getAnalyticsJsUrl';
-export * from './getAssetClickEvent';
-export * from './getRecipientInfo';
-export * from './isPrezlyTrackingAllowed';
-export * from './stringify';
-export * from './urlParameters';
