--- modules/analytics/lib/stringify.ts
@@ -1,3 +1,0 @@
-export function stringify(...args: Array<any>) {
-    return args.map((value) => JSON.stringify(value)).join(', ');
-}
