+++ package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@hcaptcha/react-hcaptcha": "^0.3.7",
     "@headlessui/react": "^1.4.1",
+    "@prezly/analytics-nextjs": "^0.1.1",
     "@prezly/content-renderer-react-js": "^0.13.0",
     "@prezly/sdk": "^6.17.0",
     "@prezly/slate-types": "^0.19.0",
