+++ pages/_app.tsx
@@ -1,10 +1,10 @@
+import { AnalyticsContextProvider } from '@prezly/analytics-nextjs';
 import type { PageProps } from '@prezly/theme-kit-nextjs';
 import { DEFAULT_LOCALE, LocaleObject, NewsroomContextProvider } from '@prezly/theme-kit-nextjs';
 import type { AppProps } from 'next/app';
 import { useMemo } from 'react';
 import { IntlProvider } from 'react-intl';
 
-import { AnalyticsContextProvider } from '@/modules/analytics';
 import type { BasePageProps } from 'types';
 
 import '@prezly/uploadcare-image/build/styles.css';
